requirements.txt
docopt==0.6.2
docstring-parser==0.15
google-api-python-client==2.97.0
google-auth==2.22.0
importlib-resources==6.0.1
Jinja2==3.1.2
packaging==23.1
pandas==1.3.5
pipreqs==0.4.13
pydantic==2.3.0
PyYAML==6.0.1
yarg==0.1.9
mock
pylint
pytest
pytest-cov
pytest-mock
pytest-xdist